moved WorkSpaceConfig into conf.py
