+++ neo4j/work/simple.py
@@ -26,7 +26,10 @@ from time import perf_counter, sleep
 from warnings import warn
 
 from neo4j.api import READ_ACCESS, WRITE_ACCESS
-from neo4j.conf import DeprecatedAlias
+from neo4j.conf import (
+    DeprecatedAlias,
+    WorkspaceConfig,
+)
 from neo4j.data import DataHydrator, DataDehydrator
 from neo4j.exceptions import (
     Neo4jError,
@@ -36,7 +39,7 @@ from neo4j.exceptions import (
     TransactionError,
 )
 from neo4j._exceptions import BoltIncompleteCommitError
-from neo4j.work import Workspace, WorkspaceConfig
+from neo4j.work import Workspace
 from neo4j.work.summary import ResultSummary
